+++ robotiq_hande_driver/src/application.cpp
@@ -1,0 +1,32 @@
+#include "application.hpp"
+
+#include <cstdio>
+
+
+namespace hande_driver  {
+
+ApplicationLayer::ApplicationLayer()
+:   requested_position_()
+,   position_()
+,   current_()
+{
+}
+
+void ApplicationLayer::read(){
+    protocol_logic_.refresh_registers();
+
+    status_.is_reset = protocol_logic_.is_reset();
+    status_.is_ready = protocol_logic_.is_ready();
+    status_.is_moving = protocol_logic_.is_moving();
+    status_.is_stopped = protocol_logic_.is_stopped();
+    status_.is_opened = protocol_logic_.is_opened();
+    status_.is_closed = protocol_logic_.is_closed();
+    status_.object_detected = protocol_logic_.obj_detected();
+
+    //fault_status
+
+    requested_position_ = GRIPPER_POSITION_MAX - (double)protocol_logic_.get_reg_pos() * GRIPPER_POSITION_STEP;
+    position_ = GRIPPER_POSITION_MAX - (double)protocol_logic_.get_pos() * GRIPPER_POSITION_STEP;
+    current_ = (double)protocol_logic_.get_current() * GRIPPER_CURRENT_SCALE;
+}
+}   // namespace hande_driver